refactor: 채팅 입력창 MessageInput을 Container/Presentational 패턴으로 분리
