+++ src/components/chat/MessageInput.jsx
@@ -1,60 +1,14 @@
 import PropTypes from "prop-types";
-
-import { useEffect, useRef, useState } from "react";
-import { addDoc, collection, serverTimestamp } from "firebase/firestore";
-import { db } from "services/firebase";
-import { useSelector } from "react-redux";
-
 import { Send, Image as ImageIcon } from "lucide-react";
 
-export default function MessageInput({ roomId }) {
-  const [message, setMessage] = useState("");
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const user = useSelector((state) => state.user.user);
-
-  const inputRef = useRef(null);
-
-  useEffect(() => {
-    inputRef.current?.focus();
-  }, []);
-
-  if (!roomId) {
-    console.error("MessageInput: roomId가 제공되지 않았습니다.");
-    return <div className="text-red-500">채팅방 ID 오류</div>;
-  }
-
-  const handleSend = async () => {
-    if (!message.trim() || !user || isSubmitting) return;
-    setIsSubmitting(true);
-
-    try {
-      await addDoc(collection(db, "chatMessages"), {
-        type: "text",
-        roomId,
-        message: message.trim(),
-        sender: {
-          uid: user.uid,
-          displayName: user.displayName || "익명",
-          photoURL: user.photoURL || "",
-        },
-        createdAt: serverTimestamp(),
-      });
-      setMessage("");
-    } catch (error) {
-      console.error("메시지 전송 오류:", error);
-      alert("메시지 전송에 실패했습니다.");
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  const handleKeyDown = (e) => {
-    if (e.key === "Enter" && !e.shiftKey) {
-      e.preventDefault();
-      handleSend();
-    }
-  };
-
+export default function MessageInput({
+  message,
+  setMessage,
+  handleSend,
+  handleKeyDown,
+  isSubmitting,
+  inputRef,
+}) {
   return (
     <div className="flex items-center gap-3">
       <button className="text-gray-500 hover:text-gray-700">
@@ -92,5 +46,9 @@ export default function MessageInput({ roomId }) {
 }
 
 MessageInput.propTypes = {
-  roomId: PropTypes.string.isRequired,
+  message: PropTypes.string.isRequired,
+  setMessage: PropTypes.func.isRequired,
+  handleSend: PropTypes.func.isRequired,
+  handleKeyDown: PropTypes.func.isRequired,
+  isSubmitting: PropTypes.bool.isRequired,
 };
